covers and masks
layers have different "cover" types, defining how the image gets transformed to fit the full map
alpha mask is applied to layer, alpha levels are multiplied so original image can still have alpha in it, or may be small and not fill the entire screen.
cover method not applied to mask, maybe it should?